+++ engelsystem-base/src/main/kotlin/info/metadude/kotlin/library/engelsystem/EngelsystemApi.kt
@@ -1,12 +1,12 @@
 package info.metadude.kotlin.library.engelsystem
 
-import okhttp3.OkHttpClient
+import okhttp3.Call
 
 interface EngelsystemApi {
 
     fun provideEngelsystemService(
         baseUrl: String,
-        okHttpClient: OkHttpClient
+        callFactory: Call.Factory,
     ): EngelsystemService
 
 }